refactor service discovery keepaliver
